+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - [command_bundle] feature: add `enable` property for each `bundle command`, making it easier to configure.
 - [core] feature: fine-tune the `command registration point` for better compatibility, include: `command_alias`, `command_bundle`, `command_permission` modules.
 - [core] feature: now will emit a console warning if a `command registration overriding` is detected.
+- [command_event] feature: added 2 server lifecycle events.
+  - feature: add `on_server_started` event.
+  - feature: add `on_server_stopping` event.
 
 # Fixed
 - [command_bundle && command_alias] fix: the `hot-reload` feature should `un-register` the target command.
+